# crates/ene-api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ene-core API",
    "version": "1.0.0",
    "description": "HTTP + WebSocket surface of the dialogue lane. Same contract for desktop, CLI, and Web."
  },
  "servers": [{ "url": "/api/v1" }],
  "components": {
    "securitySchemes": {
      "bearer": { "type": "http", "scheme": "bearer" }
    }
  },
  "security": [{ "bearer": [] }],
  "paths": {
...
    "/sessions/{id}/messages": { "post": { "summary": "prompt | steer | follow_up", "responses": { "200": { "description": "accepted" }, "409": { "description": "lane_busy" } } } },
    "/sessions/{id}/history": { "get": { "summary": "Projected history", "parameters": [{ "name": "depth", "in": "query", "schema": { "enum": ["surface", "detail"] } }], "responses": { "200": { "description": "history" } } } },
...
    "/approvals/{id}/respond": { "post": { "summary": "First writer wins", "responses": { "200": { "description": "ok" }, "409": { "description": "already_resolved" } } } },
...
    "/exclusive/{resource}": {
      "post": { "summary": "Claim mic|speaker|notify", "responses": { "200": { "description": "snapshot" }, "409": { "description": "busy" } } },
      "delete": { "summary": "Release", "responses": { "200": { "description": "snapshot" } } }
    },
...
  }
}